requirements.txt
Pillow==9.5.0
py-netgames-client == 0.0.5
